% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictExample.R
\name{predictExample}
\alias{predictExample}
\title{Run PRIMEmodel on ctss_rse_chr16to17.rds from extdata}
\usage{
predictExample(python_path = NULL, log_dir = NULL, keep_tmp = FALSE, ...)
}
\arguments{
\item{python_path}{Path to Python virtual environment}

\item{outdir}{Output directory for the pipeline}

\item{num_cores}{Number of CPU cores to use}
}
\value{
GRanges or GRangesList
}
\description{
Run PRIMEmodel on ctss_rse_chr16to17.rds from extdata
}
