% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictFocalExample.R
\name{predictFocalExample}
\alias{predictFocalExample}
\title{Run an Example of the PRIMEmodel focal Pipeline}
\usage{
predictFocalExample(python_path = NULL, log_dir = NULL, keep_tmp = FALSE, ...)
}
\arguments{
\item{python_path}{Character path to the Python binary
in the desired environment. Default is NULL.}

\item{log_dir}{Optional character path to save the log file.}

\item{keep_tmp}{Logical. If `TRUE`, temporary files and folders are retained.
Default is `FALSE`.}
}
\value{
A `GRanges` object or a `GRangesList` object,
  depending on the number of samples in the example dataset.
}
\description{
This function demonstrates how to run
the `PRIMEmodel::predictFocal()` pipeline
using example CTSS and region data bundled with the `PRIMEmodel` package.
It loads pre-packaged test data, runs the pipeline, and returns the result.
}
